v3.2
Code tweaks and bug fixes.
- Added a link to ‘settime’ from the DungeonTimers class to /expugnfree
- Removed the ‘List’ data under timers.yml, timetrials.yml, and
warps.yml. The code was changed to not require that data to be used.
- timers.yml is cleaned on every execute of /expugn. Any dungeon timer
entry that is overdue will be deleted to save disk space.
- marriage.yml is cleaned on every execute of /marriage. Any player
data that is empty will be deleted to save disk space. Also removed
checkCooldown() and added it to the cleanConfig() function. Cooldowns
and effectcooldowns that are overdue will be removed as well to save
disk space.