v3.4 - ItemDrop, ConsoleCommand, and More
- Added a new command: ‘/expugnconsole’. This command is only usable
via console and will be updated with more console only commands in the
future. Right now only ‘runitemdrop’ is attached to expugnconsole.
- Bumped ExpugnCommand to v2.2
- ‘/expugnfree’ now uses a config file to store the words used in
‘/expugnfree name [playername]’. View the attached ‘extras.yml’ for a
default configuration file.
- New feature: ItemDrop. I really liked the whole item scatter I made
for Gollem so I made it into a feature. You can make ItemSets that hold
a list of items. If you use ItemDrop, it will take a ItemSet that you
made and randomly throw items from that ItemSet around in a puff of
smoke and particles.